requirements.txt
together
fastapi[standard]
python-dotenv
uvicorn